+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 2.0.0
 ------------------
-- Support for multiline content.
+- Multiline content is supported by beginning lines with `|`.
+- Static attributes are now denoted with a leading single quote, such as `'msg Hello!` rather than a double quote.
+- Self-closing elements now have proper trailing newline when `keepWhitespace` is enabled.
+- Unanchored attributes (attributes whose indentation don't match a preceding tag) are now properly discarded.
+- Non-Louk template section languages are now properly respected.
 
 1.0.2
 ------------------
-- Always use the first attribute value declared for an element.
+- The first attribute value declared for an element is now always used.
 
 1.0.1
 ------------------
-- Properly handle lines that consist only of whitespace characters.
+- Lines that consist only of whitespace characters are properly handled.